Stabilization and Simplification of `footnote(legacy)`
